Refactor Dashboard component
