--- app/components/dashboard_component/dashboard_component.html.slim
@@ -1,16 +1,0 @@
-= tag.div(class: classes, **html_attributes) do
-  .govuk-width-container
-    - if heading
-      .dashboard-component__header
-        h2.govuk-heading-m class="govuk-!-margin-bottom-2"
-          = heading
-
-        - if link.present?
-          .dashboard-component__button = govuk_button_link_to link[:text], link[:url], class: "govuk-!-margin-bottom-0"
-
-    .dashboard-component-navigation
-      .dashboard-component-navigation__container
-        .dashboard-component-navigation__nav aria-label="Dashboard navigation"
-          ul.dashboard-component-navigation__list
-            - navigation_items.each do |item|
-              = item
