+++ app/components/publishers/no_vacancies_component/no_vacancies_component.html.slim
@@ -1,6 +1,8 @@
-= render(DashboardComponent.new(heading: t("jobs.manage.heading_html", organisation: @organisation.name, email: @email)))
+.govuk-grid-row class="govuk-!-margin-top-5"
+  .govuk-grid-column-full
+    h2.govuk-heading-m = t("jobs.manage.heading_html", organisation: @organisation.name, email: @email)
 
-.govuk-grid-row.no-jobs-border class="govuk-!-margin-top-5 govuk-!-margin-bottom-5"
+.govuk-grid-row.no-jobs-border class="govuk-!-margin-bottom-5"
   .govuk-grid-column-full
     h2.govuk-heading-m = t(".heading")
