+++ app/components/tabs_component/tabs_component.html.slim
@@ -1,0 +1,4 @@
+= tag.div(class: classes, **html_attributes) do
+  ul.tabs-component-navigation__list
+    - navigation_items.each do |item|
+      = item
