+++ app/views/publishers/vacancies/_tabs_show.html.slim
@@ -1,10 +1,10 @@
 .govuk-grid-row
   .govuk-grid-column-full
-    = render(DashboardComponent.new) do |dashboard|
-      - dashboard.navigation_item item: (govuk_link_to t("buttons.manage_listing"), organisation_job_path(vacancy.id), class: "dashboard-component-navigation__link", "aria-current": ("page" if controller_name == "vacancies"))
+    = tabs do |tabs|
+      - tabs.navigation_item item: (govuk_link_to t("buttons.manage_listing"), organisation_job_path(vacancy.id), class: "tabs-component-navigation__link", "aria-current": ("page" if controller_name == "vacancies"))
       - if vacancy.can_receive_job_applications? && !current_organisation.local_authority?
         - if vacancy.within_data_access_period?
-          - dashboard.navigation_item item: (govuk_link_to t(".applications.show_count", count: vacancy.job_applications.where(status: %w[submitted reviewed shortlisted unsuccessful]).count), organisation_job_job_applications_path(vacancy.id), class: "dashboard-component-navigation__link", "aria-current": ("page" if controller_name == "job_applications"))
+          - tabs.navigation_item item: (govuk_link_to t(".applications.show_count", count: vacancy.job_applications.where(status: %w[submitted reviewed shortlisted unsuccessful]).count), organisation_job_job_applications_path(vacancy.id), class: "tabs-component-navigation__link", "aria-current": ("page" if controller_name == "job_applications"))
         - else
-          - dashboard.navigation_item item: (govuk_link_to t(".applications.hide_count"), organisation_job_job_applications_path(vacancy.id), class: "dashboard-component-navigation__link", "aria-current": ("page" if controller_name == "job_applications"))
-      - dashboard.navigation_item item: (govuk_link_to t(".statistics"), organisation_job_statistics_path(vacancy.id), class: "dashboard-component-navigation__link", "aria-current": ("page" if controller_name == "statistics"))
+          - tabs.navigation_item item: (govuk_link_to t(".applications.hide_count"), organisation_job_job_applications_path(vacancy.id), class: "tabs-component-navigation__link", "aria-current": ("page" if controller_name == "job_applications"))
+      - tabs.navigation_item item: (govuk_link_to t(".statistics"), organisation_job_statistics_path(vacancy.id), class: "tabs-component-navigation__link", "aria-current": ("page" if controller_name == "statistics"))
