+++ app/views/publishers/vacancies/statistics/show.html.slim
@@ -1,6 +1,7 @@
 - content_for :page_title_prefix, t(".title")
 
-= render "publishers/vacancies/banner", vacancy: vacancy, job_applications: vacancy.job_applications
+= render "publishers/vacancies/banner", vacancy: vacancy
+= render "publishers/vacancies/tabs_show", vacancy: vacancy
 
 .govuk-grid-row.vacancy class="govuk-!-margin-top-7"
   - if vacancy.draft? || vacancy.pending?
