+++ spec/system/publishers_can_see_the_vacancies_dashboard_spec.rb
@@ -53,7 +53,7 @@ RSpec.describe "Publishers can see the vacancies dashboard" do
     scenario "with published vacancies" do
       visit organisation_path
 
-      within(".dashboard-component-navigation__list") do
+      within(".tabs-component-navigation__list") do
         click_on(I18n.t("publishers.vacancies_component.published.tab_heading"))
       end
 
@@ -66,7 +66,7 @@ RSpec.describe "Publishers can see the vacancies dashboard" do
     scenario "with draft vacancies" do
       visit organisation_path
 
-      within(".dashboard-component-navigation__list") do
+      within(".tabs-component-navigation__list") do
         click_on(I18n.t("publishers.vacancies_component.draft.tab_heading"))
       end
 
@@ -82,7 +82,7 @@ RSpec.describe "Publishers can see the vacancies dashboard" do
     scenario "with pending vacancies" do
       visit organisation_path
 
-      within(".dashboard-component-navigation__list") do
+      within(".tabs-component-navigation__list") do
         click_on(I18n.t("publishers.vacancies_component.pending.tab_heading"))
       end
 
@@ -98,7 +98,7 @@ RSpec.describe "Publishers can see the vacancies dashboard" do
     scenario "with expired vacancies" do
       visit organisation_path
 
-      within(".dashboard-component-navigation__list") do
+      within(".tabs-component-navigation__list") do
         click_on(I18n.t("publishers.vacancies_component.expired.tab_heading"))
       end
 
@@ -117,7 +117,7 @@ RSpec.describe "Publishers can see the vacancies dashboard" do
         draft_vacancy
         visit organisation_path
 
-        within(".dashboard-component-navigation__list") do
+        within(".tabs-component-navigation__list") do
           click_on(I18n.t("publishers.vacancies_component.draft.tab_heading"))
         end
